requirements.txt
feedparser==6.0.11
httpx==0.27.0
beautifulsoup4==4.12.3
google-generativeai==0.4.1
supabase==2.4.3
python-dotenv==1.0.1
pydantic==2.6.4
typing-extensions==4.10.0
redis==5.0.1
